SNS -> SQS rename
