+++ templates/content/SqsEventFunction/Notification.cs
@@ -1,9 +1,9 @@
 using Newtonsoft.Json;
 
-namespace SnsEventFunction
+namespace SqsEventFunction
 {
     /*
-        This class represents the notification you push into SNS and that is forwarded to this Lambda
+        This class represents the message you push into SQS and that is forwarded to this Lambda
         Add properties so that the message can be properly deserialized.
     */
     public class Notification
