+++ templates/content/SqsEventFunction/SqsEventFunction.csproj
@@ -1,13 +1,13 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp1.0</TargetFramework>
+    <TargetFramework>netcoreapp2.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="Amazon.Lambda.Core" Version="1.0.0" />
     <PackageReference Include="Amazon.Lambda.Serialization.Json" Version="1.2.0" />
-    <PackageReference Include="Kralizek.Lambda.Template.Sns" Version="2.0.0" />
+    <PackageReference Include="Kralizek.Lambda.Template.Sqs" Version="2.0.0" />
   </ItemGroup>
 
   <ItemGroup>
