+++ templates/content/SqsEventFunction/aws-lambda-tools-defaults.json
@@ -12,11 +12,11 @@
   "region": "",
   "configuration": "Release",
   "framework": "netcoreapp2.0",
-  "function-name": "SnsEventFunction",
+  "function-name": "SqsEventFunction",
   "function-role": "",
   "function-runtime": "dotnetcore2.0",
   "function-memory-size": 128,
   "function-timeout": 30,
-  "function-handler": "SnsEventFunction::SnsEventFunction.Function::FunctionHandlerAsync",
+  "function-handler": "SqsEventFunction::SqsEventFunction.Function::FunctionHandlerAsync",
   "environment-variables": "Environment=;"
 }
